requirements.txt
PyYAML==3.10
beautifulsoup4==4.1.3
cssselect==0.8
dash.py==0.1
distribute==0.6.34
doc2dash==1.1.0
docopt==0.6.1
lxml==3.0.2
parguments==0.3.1
pyquery==1.2.4
requests==1.2.3
wsgiref==0.1.2
